third_party/WebKit/LayoutTests/inspector/tracing/timeline-paint-with-layout-invalidations-on-deleted-node-expected.txt

Tests the Timeline API instrumentation of layout invalidations on a deleted node.



Running: testLocalFrame

Running: testSubframe
